separate damage.utils functions into new module
