+++ connectionist/damage/shrink_layer.py
@@ -1,15 +1,8 @@
-from typing import List, Callable, Tuple, Union
-import random
+from typing import List, Callable
 from dataclasses import dataclass
+import random
 import tensorflow as tf
-
-
-def check_shapes(donor: tf.keras.Model, recipient: tf.keras.Model) -> None:
-    """Check that the shapes of the weights are the same."""
-    for w_donor, w_recipient in zip(donor.weights, recipient.weights):
-        print(
-            f"Checking: {w_donor.name}: {w_donor.shape} -> {w_recipient.name} : {w_recipient.shape}, shape changed: {w_donor.shape == w_recipient.shape}"
-        )
+from .utils import get_weights, copy_transplant
 
 
 @dataclass
@@ -61,29 +54,6 @@ def make_recipient(model, surgery_plan: SurgeryPlan, make_model_fn: Callable):
     config[to_config_key[surgery_plan.layer]] = surgery_plan.keep_n
     print(f"New config: {config}")
     return make_model_fn(**config)
-
-
-def get_weights(model: tf.keras.Model, weight_name: str) -> tf.Tensor:
-    """Get weights by full name or abbreviation."""
-
-    # Full name matching
-    matched = [w for w in model.weights if w.name == weight_name]
-
-    # Fall back to abbreviation matching of full match failed
-    if len(matched) == 0:
-        matched = [
-            w
-            for w in model.weights
-            if model.weights_abbreviations[weight_name] in w.name
-        ]
-
-    if len(matched) == 0:
-        raise ValueError(f"Could not find weights for {weight_name}")
-    if len(matched) > 1:
-        raise ValueError(
-            f"Found multiple weights: {[w.name for w in matched]} with similar name to {weight_name}, please be more specific."
-        )
-    return matched[0]
 
 
 class Surgeon:
@@ -166,24 +136,6 @@ class Surgeon:
             )  # slice the weight in each connected axis
         w_recipient.assign(w)  # assign the weights to recipient
 
-    def copy_transplant(
-        self, donor: tf.keras.Model, recipient: tf.keras.Model, weight_name: str
-    ) -> None:
-        """Transplant weights from donor to recipient in weights that has matching shape."""
-
-        w_recipient = get_weights(recipient, weight_name)
-        w_donor = get_weights(donor, weight_name)
-
-        print(
-            f"Transplanting: {w_donor.name}:{w_donor.shape} -> {w_recipient.name}: {w_recipient.shape}"
-        )
-
-        assert (
-            w_donor.shape == w_recipient.shape
-        ), f"Shapes don't match: {w_donor.shape} != {w_recipient.shape}"
-
-        w_recipient.assign(w_donor)
-
     def transplant(self, donor: tf.keras.Model, recipient: tf.keras.Model) -> None:
         """Transplant all the weights from donor to recipient model."""
 
@@ -207,4 +159,4 @@ class Surgeon:
             name for name in all_weights_names if name not in weight_names
         ]
         for name in remaining_weights:
-            self.copy_transplant(donor=donor, recipient=recipient, weight_name=name)
+            copy_transplant(donor=donor, recipient=recipient, weight_name=name)
